requirements.txt
# Core dependencies
requests>=2.25.1
beautifulsoup4>=4.9.3
dnspython>=2.1.0
docker>=5.0.3
pandas>=1.3.0
openpyxl>=3.0.7
python-dotenv>=0.19.0

# API integrations
anthropic>=0.3.0
openai>=0.27.0

# Utilities
tqdm>=4.61.2
colorama>=0.4.4
retry>=0.9.2
tenacity>=8.0.1
urllib3>=1.26.6
lxml>=4.6.3